Hardening pass + design-doc update (post-M2 review)
Robustness fixes for input that crosses the R<->Rust boundary, plus the
deferred findings recorded in DESIGN.md.

Rust (prevent panics that abort the R session):
- rs_scene validates width/height/dpi are finite and positive and caps
  pixel dimensions (NaN/Inf/huge previously panicked Pixmap::new or
  attempted a runaway allocation)
- rgba() capacity computed in usize (u32 product could overflow)
- per-glyph text vectors length-clamped before indexing
- glyph ids carried as u32 end-to-end (no u16 truncation -> wrong glyph)
- parenthesize the rotation angle cast for clarity

R:
- rs_lines/rs_polygon error on mismatched x/y length (like rs_bbox)
- gpar colour/number args enforce length-1 (a length>1 colour silently
  became "no paint" before)

Tests: add test-validation.R (errors, NA=none vs NULL=inherit, col/lwd
inheritance) -- coverage gaps from the review.

Docs: DESIGN.md roadmap marked M0-M2 done with as-built scope; add a
"Hardening backlog" listing deferred items (scalar->vector contracts,
broader validation, render caching, up/pop semantics) for M3.

Examples: drop a stray devtools::load_all() from scatter.R.

76 tests pass; R CMD check 0/0/0; clean-room offline install verified.

+++ src/rust/src/font.rs
@@ -36,13 +36,13 @@ impl FontCache {
         &mut self,
         font_path: &str,
         face_index: u32,
-        glyph_id: u16,
+        glyph_id: u32,
         size_px: f32,
     ) -> Option<tiny_skia::Path> {
         let bytes = self.bytes(font_path)?;
         let font = FontRef::from_index(bytes, face_index).ok()?;
         let outlines = font.outline_glyphs();
-        let glyph = outlines.get(GlyphId::new(glyph_id as u32))?;
+        let glyph = outlines.get(GlyphId::new(glyph_id))?;
 
         let mut pen = PathPen::default();
         let settings = DrawSettings::unhinted(Size::new(size_px), LocationRef::default());
